auto-bot/requirements.txt
python-telegram-bot[job-queue]
python-dotenv
pyyaml
sqlalchemy
